+++ src/ert/__init__.py
@@ -1,4 +1,5 @@
 from .data import MeasuredData
+from .libres_facade import LibresFacade
 from .simulator import BatchSimulator
 
-__all__ = ["MeasuredData", "BatchSimulator"]  # Used by semeio
+__all__ = ["MeasuredData", "LibresFacade", "BatchSimulator"]  # Used by semeio
