+++ src/ert/gui/ertwidgets/models/analysismodulevariablesmodel.py
@@ -15,7 +15,7 @@
 #  for more details.
 from typing import List
 
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from res.analysis.analysis_module import AnalysisModule
