+++ src/ert/gui/simulation/ensemble_smoother_panel.py
@@ -11,7 +11,7 @@ from ert.gui.ertwidgets.models.targetcasemodel import TargetCaseModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert_shared.ide.keywords.definitions import RangeStringArgument, ProperNameArgument
 from ert.gui.simulation import SimulationConfigPanel
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from ert_shared.models import EnsembleSmoother
 from res.enkf import EnKFMain
