+++ src/ert/gui/simulation/iterated_ensemble_smoother_panel.py
@@ -11,7 +11,7 @@ from ert_shared.ide.keywords.definitions import (
 )
 from ert.gui.simulation import SimulationConfigPanel
 from ert_shared.models import IteratedEnsembleSmoother
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 
 from dataclasses import dataclass
