+++ src/ert_shared/dark_storage/common.py
@@ -2,7 +2,7 @@ from typing import List, Union
 import pandas as pd
 
 from ert.data import loader, MeasuredData
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from res.enkf import EnkfObservationImplementationType
