+++ src/ert_shared/dark_storage/enkf.py
@@ -3,7 +3,7 @@ import os
 from typing import Optional
 from fastapi import Depends
 from ert_storage.security import security
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from res.enkf import EnKFMain, ResConfig
 
 __all__ = ["LibresFacade", "get_res"]
