+++ src/ert_shared/storage/extraction.py
@@ -14,7 +14,7 @@ import io
 
 
 if TYPE_CHECKING:
-    from ert_shared.libres_facade import LibresFacade
+    from ert.libres_facade import LibresFacade
 
 logger = logging.getLogger()
