+++ tests/ert_tests/shared/test_libres_facade.py
@@ -5,7 +5,7 @@ import shutil
 import pandas as pd
 from pandas.core.base import PandasObject
 
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from res.enkf import EnKFMain, ResConfig
 from res.enkf.export import SummaryCollector
