+++ tests/ert_tests/storage/test_extraction.py
@@ -12,7 +12,7 @@ import pandas as pd
 import pytest
 from numpy.testing import assert_almost_equal, assert_array_equal
 
-from ert_shared.libres_facade import LibresFacade
+from ert.libres_facade import LibresFacade
 from ert_shared.storage import extraction
 from res.enkf import RunContext
 from res.enkf.enkf_main import EnKFMain
